requirements.txt
# AI for Product Managers - Dependencies
# Stage 1: Core Dependencies

# UI Framework
streamlit>=1.28.0

# AutoGen (Microsoft Multi-Agent Framework)
pyautogen>=0.2.0

# Data Collection (Stage 2+)
pytrends>=4.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
praw>=7.7.0  # Reddit API

# NLP & Text Processing
scikit-learn>=1.3.0
nltk>=3.8.0

# Text Quality & Readability
textstat>=0.7.3

# PDF Generation
fpdf2>=2.7.0
reportlab>=4.0.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Date/Time handling
python-dateutil>=2.8.0
